requirements.txt
certifi==2017.11.5
Django==1.11.7
django-admin-page-lock==1.2.0
django-haystack==2.6.1
django-lock-tokens==0.1.4
django-tinymce==2.6.0
elasticsearch==1.9.0
elasticsearch2==2.5.0
mysqlclient==1.3.12
pyelasticsearch==1.4
pytz==2017.3
simplejson==3.12.0
six==1.11.0
urllib3==1.22
Whoosh==2.7.4
